+++ static/src/javascripts/lib/url.js
@@ -50,9 +50,9 @@ const constructQuery = (query: Object): string =>
     Object.keys(query)
         .map(param => {
             const value = query[param];
-            const queryValue = Array.isArray(value) ?
-                value.map(v => encodeURIComponent(v)).join(',') :
-                encodeURIComponent(value);
+            const queryValue = Array.isArray(value)
+                ? value.map(v => encodeURIComponent(v)).join(',')
+                : encodeURIComponent(value);
             return `${param}=${queryValue}`;
         })
         .join('&');
