+++ static/src/javascripts/projects/common/modules/commercial/acquisitions-ophan.js
@@ -1,5 +1,7 @@
 // @flow
 import ophan from 'ophan/ng';
+import config from 'lib/config';
+import { constructQuery as constructURLQuery } from 'lib/url';
 
 type ComponentEventWithoutAction = {
     component: OphanComponent,
@@ -28,3 +30,29 @@ export const submitViewEvent = (componentEvent: ComponentEventWithoutAction) =>
         ...componentEvent,
         action: 'VIEW',
     });
+
+export const addTrackingCodesToUrl = (
+    base: string,
+    componentType: OphanComponentType,
+    campaignCode: string,
+    abTest?: { name: string, variant: string }
+) => {
+    const acquisitionData = {
+        source: 'GUARDIAN_WEB',
+        componentId: campaignCode,
+        componentType,
+        referrerPageviewId: config.get('ophan.pageViewId') || undefined,
+        campaignCode,
+        abTest,
+    };
+
+    const params = {
+        REFPVID: config.get('ophan.pageViewId') || 'not_found',
+        INTCMP: campaignCode,
+        acquisitionData: JSON.stringify(acquisitionData),
+    };
+
+    return `${base}${base.includes('?') ? '&' : '?'}${constructURLQuery(
+        params
+    )}`;
+};
